Moved motors from sampleview to beamline
