+++ mxcube3/ui/actions/beamline.js
@@ -9,97 +9,6 @@ export const STATE = {
 };
 
 
-/**
- *  Initial redux state for beamline attributes, object containing each beamline
- *  attribute (name, attribute object). Each attribute object in turn have the
- *  follwoing properties:
- *
- *     name:   name of beamline attribute
- *     value:  attributes current value
- *     state:  attributes current state, see STATE for more information
- *     msg:    arbitray message describing current state
- */
-export const INITIAL_STATE = {
-  energy: {
-    limits: [
-      0,
-      1000,
-      0.1
-    ],
-    name: 'energy',
-    value: '0',
-    state: STATE.IDLE,
-    msg: ''
-  },
-  resolution: {
-    limits: [
-      0,
-      1000,
-      0.1
-    ],
-    name: 'resolution',
-    value: '0',
-    state: STATE.IDLE,
-    msg: ''
-  },
-  transmission: {
-    limits: [
-      0,
-      1000,
-      0.1
-    ],
-    name: 'transmission',
-    value: '0',
-    state: STATE.IDLE,
-    msg: ''
-  },
-  fast_shutter: {
-    limits: [
-      0,
-      1,
-      1
-    ],
-    name: 'fast_shutter',
-    value: 'undefined',
-    state: 'undefined',
-    msg: 'UNKNOWN'
-  },
-  safety_shutter: {
-    limits: [
-      0,
-      1,
-      1
-    ],
-    name: 'safety_shutter',
-    value: 'undefined',
-    state: 'undefined',
-    msg: 'UNKNOWN'
-  },
-  beamstop: {
-    limits: [
-      0,
-      1,
-      1
-    ],
-    name: 'beamstop',
-    value: 'undefined',
-    state: 'undefined',
-    msg: 'UNKNOWN'
-  },
-  capillary: {
-    limits: [
-      0,
-      1,
-      1
-    ],
-    name: 'capillary',
-    value: 'undefined',
-    state: 'undefined',
-    msg: 'UNKNOWN'
-  }
-};
-
-
 // Action types
 export const BL_ATTR_SET = 'BL_ATTR_SET';
 export const BL_ATTR_GET_ALL = 'BL_ATTR_GET_ALL';
