+++ mxcube3/ui/reducers/sampleview.js
@@ -18,20 +18,6 @@ const initialState = {
     kappaStep: 0.1,
     kappaphiStep: 0.1
   },
-  motors: {
-    focus: { position: 0, Status: 0 },
-    phi: { position: 0, Status: 0 },
-    phiy: { position: 0, Status: 0 },
-    phiz: { position: 0, Status: 0 },
-    sampx: { position: 0, Status: 0 },
-    sampy: { position: 0, Status: 0 },
-    BackLight: { position: 0, Status: 0 },
-    FrontLight: { position: 0, Status: 0 },
-    BackLightSwitch: { position: 0, Status: 0 },
-    FrontLightSwitch: { position: 0, Status: 0 },
-    kappa: { position: 0, Status: 0 },
-    kappa_phi: { position: 0, Status: 0 }
-  },
   pixelsPerMm: 0,
   imageRatio: 0,
   contextMenu: { show: false, shape: { type: 'NONE' }, x: 0, y: 0 },
@@ -107,36 +93,6 @@ export default (state = initialState, action) => {
           pixelsPerMm: action.pixelsPerMm
         };
       }
-    case 'SET_MOTOR_MOVING':
-      {
-        return {
-          ...state,
-          motors: { ...state.motors,
-            [action.name.toLowerCase()]: {
-              ...state.motors[action.name.toLowerCase()],
-              Status: action.status
-            }
-          }
-        };
-      }
-    case 'SAVE_MOTOR_POSITIONS':
-      {
-        return {
-          ...state,
-          motors: { ...state.motors, ...action.data },
-          zoom: action.data.zoom.position,
-          pixelsPerMm: action.data.pixelsPerMm[0]
-        };
-      }
-    case 'SAVE_MOTOR_POSITION':
-      {
-        return {
-          ...state,
-          motors: { ...state.motors,
-            [action.name]: { position: action.value }
-          }
-        };
-      }
     case 'UPDATE_POINTS_POSITION':
       {
         return { ...state, points: action.points };
@@ -185,11 +141,17 @@ export default (state = initialState, action) => {
       {
         return { ...state, motorSteps: { ...state.motorSteps, [action.name]: action.value } };
       }
+    case 'SAVE_MOTOR_POSITIONS':
+      {
+        return { ...state,
+                 zoom: action.data.zoom.position,
+                 pixelsPerMm: action.data.pixelsPerMm[0]
+               };
+      }
     case 'SET_INITIAL_STATUS':
       {
         return {
           ...state,
-          motors: { ...state.motors, ...action.data.Motors },
           zoom: action.data.Motors.zoom.position,
           width: action.data.Camera.imageWidth,
           height: action.data.Camera.imageHeight,
